[Admin] Extracted ProductOptionContext in setup
