+++ etc/behat/services/contexts/setup.xml
@@ -22,6 +22,7 @@
         <parameter key="sylius.behat.context.setup.payment.class">Sylius\Behat\Context\Setup\PaymentContext</parameter>
         <parameter key="sylius.behat.context.setup.product.class">Sylius\Behat\Context\Setup\ProductContext</parameter>
         <parameter key="sylius.behat.context.setup.product_attribute.class">Sylius\Behat\Context\Setup\ProductAttributeContext</parameter>
+        <parameter key="sylius.behat.context.setup.product_option.class">Sylius\Behat\Context\Setup\ProductOptionContext</parameter>
         <parameter key="sylius.behat.context.setup.product_review.class">Sylius\Behat\Context\Setup\ProductReviewContext</parameter>
         <parameter key="sylius.behat.context.setup.promotion.class">Sylius\Behat\Context\Setup\PromotionContext</parameter>
         <parameter key="sylius.behat.context.setup.security.class">Sylius\Behat\Context\Setup\SecurityContext</parameter>
@@ -104,10 +105,7 @@
         <service id="sylius.behat.context.setup.product" class="%sylius.behat.context.setup.product.class%" scope="scenario">
             <argument type="service" id="sylius.behat.shared_storage" container="symfony" />
             <argument type="service" id="sylius.repository.product" container="symfony" />
-            <argument type="service" id="sylius.repository.product_option" container="symfony" />
             <argument type="service" id="sylius.factory.product" container="symfony" />
-            <argument type="service" id="sylius.factory.product_option" container="symfony" />
-            <argument type="service" id="sylius.factory.product_option_value" container="symfony" />
             <argument type="service" id="sylius.factory.product_variant" container="symfony" />
             <argument type="service" id="doctrine.orm.entity_manager" container="symfony" />
             <tag name="sylius.behat.context" />
@@ -117,6 +115,15 @@
             <argument type="service" id="sylius.behat.shared_storage" container="symfony" />
             <argument type="service" id="sylius.repository.product_attribute" container="symfony" />
             <argument type="service" id="sylius.factory.product_attribute" container="symfony" />
+            <tag name="sylius.behat.context" />
+        </service>
+
+        <service id="sylius.behat.context.setup.product_option" class="%sylius.behat.context.setup.product_option.class%" scope="scenario">
+            <argument type="service" id="sylius.behat.shared_storage" container="symfony" />
+            <argument type="service" id="sylius.repository.product_option" container="symfony" />
+            <argument type="service" id="sylius.factory.product_option" container="symfony" />
+            <argument type="service" id="sylius.factory.product_option_value" container="symfony" />
+            <argument type="service" id="doctrine.orm.entity_manager" container="symfony" />
             <tag name="sylius.behat.context" />
         </service>
