+++ etc/behat/suites/ui/product/managing_product_options.yml
@@ -13,7 +13,7 @@ default:
                 - sylius.behat.context.transform.shared_storage
 
                 - sylius.behat.context.setup.channel
-                - sylius.behat.context.setup.product
+                - sylius.behat.context.setup.product_option
                 - sylius.behat.context.setup.security
 
                 - sylius.behat.context.ui.admin.managing_product_options
