+++ src/Sylius/Behat/Context/Setup/ProductContext.php
@@ -19,17 +19,13 @@ use Sylius\Component\Core\Model\ProductInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
 use Sylius\Component\Core\Repository\ProductRepositoryInterface;
 use Sylius\Component\Core\Test\Services\SharedStorageInterface;
-use Sylius\Component\Product\Model\OptionInterface;
-use Sylius\Component\Product\Model\OptionValueInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Sylius\Component\Taxation\Model\TaxCategoryInterface;
-use Sylius\Component\Variation\Repository\OptionRepositoryInterface;
 
 /**
  * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
  * @author Magdalena Banasiak <magdalena.banasiak@lakion.com>
- * @author Grzegorz Sadowski <grzegorz.sadowski@lakion.com>
  */
 final class ProductContext implements Context
 {
@@ -44,24 +40,9 @@ final class ProductContext implements Context
     private $productRepository;
 
     /**
-     * @var OptionRepositoryInterface
-     */
-    private $productOptionRepository;
-
-    /**
      * @var FactoryInterface
      */
     private $productFactory;
-
-    /**
-     * @var FactoryInterface
-     */
-    private $productOptionFactory;
-
-    /**
-     * @var FactoryInterface
-     */
-    private $productOptionValueFactory;
 
     /**
      * @var FactoryInterface
@@ -76,29 +57,20 @@ final class ProductContext implements Context
     /**
      * @param SharedStorageInterface $sharedStorage
      * @param ProductRepositoryInterface $productRepository
-     * @param OptionRepositoryInterface $productOptionRepository
      * @param FactoryInterface $productFactory
-     * @param FactoryInterface $productOptionFactory
-     * @param FactoryInterface $productOptionValueFactory
      * @param FactoryInterface $productVariantFactory
      * @param ObjectManager $objectManager
      */
     public function __construct(
         SharedStorageInterface $sharedStorage,
         ProductRepositoryInterface $productRepository,
-        OptionRepositoryInterface $productOptionRepository,
         FactoryInterface $productFactory,
-        FactoryInterface $productOptionFactory,
-        FactoryInterface $productOptionValueFactory,
         FactoryInterface $productVariantFactory,
         ObjectManager $objectManager
     ) {
         $this->sharedStorage = $sharedStorage;
         $this->productRepository = $productRepository;
-        $this->productOptionRepository = $productOptionRepository;
         $this->productFactory = $productFactory;
-        $this->productOptionFactory = $productOptionFactory;
-        $this->productOptionValueFactory = $productOptionValueFactory;
         $this->productVariantFactory = $productVariantFactory;
         $this->objectManager = $objectManager;
     }
@@ -195,48 +167,6 @@ final class ProductContext implements Context
     ) {
         $productVariant->setTaxCategory($taxCategory);
         $this->objectManager->flush($productVariant);
-    }
-
-    /**
-     * @Given the store has a product option :productOptionName with a code :productOptionCode
-     */
-    public function theStoreHasAProductOptionWithACode($productOptionName, $productOptionCode)
-    {
-        $productOption = $this->productOptionFactory->createNew();
-        $productOption->setCode($productOptionCode);
-        $productOption->setName($productOptionName);
-
-        $this->sharedStorage->set('product_option', $productOption);
-        $this->productOptionRepository->add($productOption);
-    }
-
-    /**
-     * @Given /^(this product option) has(?:| also) the "([^"]+)" option value with code "([^"]+)"$/
-     */
-    public function thisProductOptionHasTheOptionValueWithCode(
-        OptionInterface $productOption,
-        $productOptionValueName,
-        $productOptionValueCode
-    ) {
-        $productOptionValue = $this->createProductOptionValue($productOptionValueName, $productOptionValueCode);
-        $productOption->addValue($productOptionValue);
-
-        $this->objectManager->flush();
-    }
-
-    /**
-     * @param string $value
-     * @param string $code
-     *
-     * @return OptionValueInterface
-     */
-    private function createProductOptionValue($value, $code)
-    {
-        $productOptionValue = $this->productOptionValueFactory->createNew();
-        $productOptionValue->setValue($value);
-        $productOptionValue->setCode($code);
-
-        return $productOptionValue;
     }
 
     /**
